Replace Profile entities with Speaker entities
The main change in the application is the replacement of all Profile entities with Speaker entities. Several file deletion and creation operations occur dealing with Profile and Speaker classes, controllers and components, mostly in the admin section of the application. Additionally, a Referral class and controller are created and a Proposal class appears. The conceptually unrelated UpdateProfileInformationForm, UpdatePasswordForm and DeleteUserForm React components are removed. The type video.d.ts is renamed to media.d.ts. The commit touches almost all parts of the application.